Refactoring class objects to their own python file so my IDE stops stuttering every 2 seconds
